services/api-gateway/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
httpx==0.27.2
pydantic==2.9.2
PyJWT==2.9.0
slowapi==0.1.9
